remove useless class and code
