+++ src/main/java/cam72cam/mod/render/cutter/MeshPlaneCutter.java
@@ -6,52 +6,19 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.*;
 
-/**
- * Utility for cutting a list of primitives (e.g., BakedQuads) with a plane.
- * <p>
- * The algorithm:
- * <ol>
- *   <li>Each primitive is clipped against the plane using {@link PolygonClipper}.</li>
- *   <li>The resulting fragments are converted back to primitives via the adapter.</li>
- *   <li>Intersection points (pairs of entry/exit) are collected and used to build a graph.</li>
- *   <li>All closed loops (Eulerian cycles) are extracted using Hierholzer's algorithm.</li>
- *   <li>Loops are split into simple polygons if they contain repeated coordinates
- *       (which can happen when multiple loops share vertices or edges).</li>
- *   <li>Each simple loop is used to generate a cap polygon, which is then rendered.</li>
- * </ol>
- * <p>
- * The implementation handles:
- * <ul>
- *   <li>Convex polygons (direct cap generation).</li>
- *   <li>Concave polygons (ear‑clipping triangulation before cap generation).</li>
- *   <li>Multiple disjoint rings (e.g., stair treads).</li>
- *   <li>Rings sharing vertices (e.g., two rectangles touching at a corner).</li>
- *   <li>Rings sharing edges (e.g., "日"‑shaped two rectangles sharing a side).</li>
- * </ul>
- * <p>
- * All operations are compatible with Java 8.
- *
- * @see PrimitiveAdapter
- * @see PolygonClipper
- * @see CapBuilder
+/*
+ * Cuts primitives with a plane and generates caps for the cross‑section.
+ * Handles convex/concave polygons, multiple disjoint rings, and rings sharing edges/vertices.
  */
 public final class MeshPlaneCutter {
 
     private static final double EPS = 1e-4;
 
-    private MeshPlaneCutter() {
-        // Private constructor to prevent instantiation.
-    }
+    private MeshPlaneCutter() {}
 
     /**
-     * Cuts the given list of primitives with the specified plane.
-     *
-     * @param primitives the input primitives (must be convertible to polygons)
-     * @param plane      the cutting plane (positive side is kept)
-     * @param adapter    the adapter for converting between primitives and polygons
-     * @param <T>        the primitive type
-     * @param <Template> the template type for UV generation
-     * @return a list of primitives resulting from the cut (including fragments and caps)
+     * Main cutting entry.
+     * Returns fragments and caps.
      */
     public static <T, Template> List<T> cut(
             List<T> primitives,
@@ -61,7 +28,6 @@ public final class MeshPlaneCutter {
         List<T> result = new ArrayList<>();
         List<Pair<ClipVertex, ClipVertex>> allPairs = new ArrayList<>();
 
-        // 1. Clip each primitive and collect intersection point pairs.
         for (T primitive : primitives) {
             Polygon polygon = adapter.toPolygon(primitive);
             ClipResult clipped = PolygonClipper.clip(polygon, plane);
@@ -78,7 +44,6 @@ public final class MeshPlaneCutter {
             }
         }
 
-        // 2. Extract all closed loops (rings) from the intersection graph.
         List<List<ClipVertex>> rings = extractRings(allPairs);
         if (!allPairs.isEmpty() && rings.isEmpty()) {
             ModCore.error("Fail to get ring with ClipVertex Pairs:");
@@ -87,14 +52,12 @@ public final class MeshPlaneCutter {
             }
         }
 
-        // 3. Generate a cap for each ring.
         Template template = adapter.createTemplate(primitives, plane);
         if (template == null) return result;
 
         for (List<ClipVertex> ring : rings) {
             if (ring.size() < 3) continue;
 
-            // Ensure counter‑clockwise winding (relative to plane normal).
             if (signedArea(ring, plane.normal) < 0) {
                 Collections.reverse(ring);
             }
@@ -102,15 +65,12 @@ public final class MeshPlaneCutter {
             boolean isConvex = isPolygonConvex(ring, plane.normal);
 
             if (isConvex) {
-                // Convex polygon: use directly.
                 Polygon capPoly = new Polygon(ring, plane.normal);
                 adapter.prepareCap(capPoly, plane, template);
                 result.addAll(adapter.fromTemplate(capPoly, template));
             } else {
-                // Concave polygon: triangulate first.
                 List<List<ClipVertex>> triangles = earClip(ring, plane.normal);
                 for (List<ClipVertex> tri : triangles) {
-                    // Convert triangle to quad by duplicating the last vertex.
                     List<ClipVertex> quadVerts = new ArrayList<>(4);
                     quadVerts.add(tri.get(0));
                     quadVerts.add(tri.get(1));
@@ -126,31 +86,16 @@ public final class MeshPlaneCutter {
         return result;
     }
 
-    // =========================================================================
-    //  Graph‑based loop extraction (Hierholzer + splitting)
-    // =========================================================================
+    // graph-based loop extraction (Hierholzer + splitting)
 
-    /**
-     * Extracts all simple closed loops from a list of intersection edges.
-     * <p>
-     * The method:
-     * <ol>
-     *   <li>Merges vertices that are within {@link #EPS} of each other.</li>
-     *   <li>Builds a multigraph using {@code List} adjacency to preserve parallel edges.</li>
-     *   <li>For each connected component, checks that all vertex degrees are even
-     *       (Eulerian condition).</li>
-     *   <li>Repeatedly extracts Eulerian circuits using Hierholzer's algorithm.</li>
-     *   <li>Each extracted circuit is split into simple polygons if it contains
-     *       repeated coordinates (which indicate multiple loops sharing vertices or edges).</li>
-     * </ol>
-     *
-     * @param pairs the list of directed edges (entry → exit)
-     * @return a list of rings, each represented as a list of vertices in order
+    /*
+     * Extracts all simple closed loops from intersection edges.
+     * Merges vertices within EPS, builds multigraph, extracts Eulerian circuits,
+     * splits circuits with repeated coordinates into separate rings.
      */
     private static List<List<ClipVertex>> extractRings(List<Pair<ClipVertex, ClipVertex>> pairs) {
         if (pairs.isEmpty()) return Collections.emptyList();
 
-        // 1. Merge vertices with identical coordinates.
         Map<Vec3d, ClipVertex> coordMap = new HashMap<>();
         java.util.function.Function<ClipVertex, ClipVertex> getMerged = (v) -> {
             for (Vec3d key : coordMap.keySet()) {
@@ -162,7 +107,6 @@ public final class MeshPlaneCutter {
             return v;
         };
 
-        // 2. Build adjacency list (using List to allow parallel edges).
         Map<ClipVertex, List<ClipVertex>> graph = new HashMap<>();
         for (Pair<ClipVertex, ClipVertex> p : pairs) {
             ClipVertex a = getMerged.apply(p.getLeft());
@@ -177,7 +121,6 @@ public final class MeshPlaneCutter {
         for (ClipVertex start : graph.keySet()) {
             if (globalVisited.contains(start)) continue;
 
-            // 3. Extract the connected component.
             Set<ClipVertex> component = new HashSet<>();
             Deque<ClipVertex> stack = new ArrayDeque<>();
             stack.push(start);
@@ -192,7 +135,6 @@ public final class MeshPlaneCutter {
                 }
             }
 
-            // 4. Check that all degrees are even.
             boolean allEven = true;
             for (ClipVertex v : component) {
                 if (graph.get(v).size() % 2 != 0) {
@@ -205,13 +147,11 @@ public final class MeshPlaneCutter {
                 continue;
             }
 
-            // 5. Copy adjacency for modification.
             Map<ClipVertex, List<ClipVertex>> graphCopy = new HashMap<>();
             for (ClipVertex v : component) {
                 graphCopy.put(v, new ArrayList<>(graph.get(v)));
             }
 
-            // 6. Repeatedly extract Eulerian circuits.
             while (true) {
                 ClipVertex startVertex = null;
                 for (ClipVertex v : component) {
@@ -229,8 +169,7 @@ public final class MeshPlaneCutter {
                 while (!stack2.isEmpty()) {
                     if (!graphCopy.get(current).isEmpty()) {
                         stack2.push(current);
-                        ClipVertex next = graphCopy.get(current).remove(0); // remove first edge
-                        // Remove reverse edge (parallel edges remain).
+                        ClipVertex next = graphCopy.get(current).remove(0);
                         graphCopy.get(next).remove(current);
                         current = next;
                     } else {
@@ -240,13 +179,11 @@ public final class MeshPlaneCutter {
                 }
                 Collections.reverse(ring);
 
-                // Remove duplicate start point (if any).
                 if (!ring.isEmpty() && ring.get(0) == ring.get(ring.size() - 1)) {
                     ring.remove(ring.size() - 1);
                 }
 
                 if (ring.size() >= 3) {
-                    // Split if the circuit contains repeated coordinates.
                     allRings.addAll(splitRing(ring));
                 }
             }
@@ -257,20 +194,9 @@ public final class MeshPlaneCutter {
         return allRings;
     }
 
-    // =========================================================================
-    //  Ring splitting (based on repeated coordinates)
-    // =========================================================================
-
-    /**
-     * Recursively splits a ring that may contain repeated vertices (due to shared
-     * vertices or edges) into simple polygons with no repeated coordinates.
-     * <p>
-     * The method finds a coordinate that appears at least twice, picks a split
-     * pair that gives the largest gap (to avoid tiny rings), and cuts the ring
-     * at those positions. The resulting sub‑rings are processed recursively.
-     *
-     * @param ring the input ring (may have repeated coordinates)
-     * @return a list of simple rings (no repeated coordinates) with at least 3 vertices
+    /*
+     * Splits a ring with repeated coordinates into simple sub‑rings.
+     * Recursively cuts at the largest gap between duplicate coordinates.
      */
     private static List<List<ClipVertex>> splitRing(List<ClipVertex> ring) {
         List<List<ClipVertex>> result = new ArrayList<>();
@@ -279,14 +205,12 @@ public final class MeshPlaneCutter {
             return result;
         }
 
-        // 1. Collect indices for each coordinate.
         Map<Vec3d, List<Integer>> posIndices = new HashMap<>();
         for (int i = 0; i < ring.size(); i++) {
             Vec3d pos = ring.get(i).pos;
             posIndices.computeIfAbsent(pos, k -> new ArrayList<>()).add(i);
         }
 
-        // 2. Find coordinates that appear at least twice.
         List<Map.Entry<Vec3d, List<Integer>>> duplicates = new ArrayList<>();
         for (Map.Entry<Vec3d, List<Integer>> entry : posIndices.entrySet()) {
             if (entry.getValue().size() >= 2) {
@@ -295,12 +219,10 @@ public final class MeshPlaneCutter {
         }
 
         if (duplicates.isEmpty()) {
-            // No duplicates: this ring is already simple.
             result.add(ring);
             return result;
         }
 
-        // 3. Pick the duplicate pair with the largest gap (first occurrence to second).
         Map.Entry<Vec3d, List<Integer>> best = null;
         int bestGap = -1;
         for (Map.Entry<Vec3d, List<Integer>> entry : duplicates) {
@@ -323,13 +245,11 @@ public final class MeshPlaneCutter {
         int first = idxs.get(0);
         int second = idxs.get(1);
 
-        // 4. Split the ring into two sub‑rings.
         List<ClipVertex> sub1 = new ArrayList<>(ring.subList(first, second + 1));
         List<ClipVertex> sub2 = new ArrayList<>();
         sub2.addAll(ring.subList(second, ring.size()));
         sub2.addAll(ring.subList(0, first + 1));
 
-        // Remove duplicate end vertices (since we included the start point in both).
         if (sub1.size() > 1 && sub1.get(0) == sub1.get(sub1.size() - 1)) {
             sub1.remove(sub1.size() - 1);
         }
@@ -337,13 +257,11 @@ public final class MeshPlaneCutter {
             sub2.remove(sub2.size() - 1);
         }
 
-        // Guard against infinite recursion: sub‑rings must be strictly smaller.
         if (sub1.size() >= ring.size() || sub2.size() >= ring.size()) {
             result.add(ring);
             return result;
         }
 
-        // 5. Recurse on each sub‑ring.
         if (sub1.size() >= 3) {
             result.addAll(splitRing(sub1));
         } else {
@@ -358,19 +276,8 @@ public final class MeshPlaneCutter {
         return result;
     }
 
-    // =========================================================================
-    //  Geometric utilities
-    // =========================================================================
+    // geometry helpers
 
-    /**
-     * Computes the signed area of a polygon projected onto the plane normal.
-     * <p>
-     * The sign indicates the winding direction (positive = counter‑clockwise).
-     *
-     * @param poly   the polygon vertices
-     * @param normal the plane normal
-     * @return the signed area
-     */
     private static double signedArea(List<ClipVertex> poly, Vec3d normal) {
         double area = 0;
         int n = poly.size();
@@ -382,16 +289,6 @@ public final class MeshPlaneCutter {
         return area;
     }
 
-    /**
-     * Checks if a polygon is convex.
-     * <p>
-     * All cross products of consecutive edges must have the same sign
-     * (relative to the plane normal).
-     *
-     * @param polygon the polygon vertices
-     * @param normal  the plane normal
-     * @return true if the polygon is convex
-     */
     private static boolean isPolygonConvex(List<ClipVertex> polygon, Vec3d normal) {
         int n = polygon.size();
         if (n < 3) return false;
@@ -413,20 +310,8 @@ public final class MeshPlaneCutter {
         return true;
     }
 
-    // =========================================================================
-    //  Ear‑clipping triangulation
-    // =========================================================================
+    // ear‑clipping triangulation
 
-    /**
-     * Triangulates a concave polygon using the ear‑clipping algorithm.
-     * <p>
-     * The polygon must be simple, planar, and have vertices in counter‑clockwise
-     * order. The result is a list of triangles, each as a list of 3 vertices.
-     *
-     * @param polygon the input polygon (must be CCW)
-     * @param normal  the plane normal (used for convexity tests)
-     * @return a list of triangles (each a list of 3 vertices)
-     */
     private static List<List<ClipVertex>> earClip(List<ClipVertex> polygon, Vec3d normal) {
         List<List<ClipVertex>> triangles = new ArrayList<>();
         List<ClipVertex> verts = new ArrayList<>(polygon);
@@ -451,7 +336,6 @@ public final class MeshPlaneCutter {
                 }
             }
             if (!found) {
-                // Fallback: take the first three vertices.
                 List<ClipVertex> fallback = new ArrayList<>(Arrays.asList(verts.get(0), verts.get(1), verts.get(2)));
                 if (signedArea(fallback, normal) < 0) {
                     Collections.reverse(fallback);
@@ -471,16 +355,6 @@ public final class MeshPlaneCutter {
         return triangles;
     }
 
-    /**
-     * Checks if the angle at vertex b is convex (ear condition).
-     * Assumes vertices are in counter‑clockwise order.
-     *
-     * @param a      previous vertex
-     * @param b      current vertex
-     * @param c      next vertex
-     * @param normal plane normal
-     * @return true if the ear is convex (cross product > 0)
-     */
     private static boolean isConvexEar(ClipVertex a, ClipVertex b, ClipVertex c, Vec3d normal) {
         Vec3d ab = b.pos.subtract(a.pos);
         Vec3d bc = c.pos.subtract(b.pos);
@@ -488,19 +362,8 @@ public final class MeshPlaneCutter {
         return cross > 0;
     }
 
-    /**
-     * Tests whether any vertex from the given list lies inside the triangle (a,b,c).
-     *
-     * @param a      triangle vertex 1
-     * @param b      triangle vertex 2
-     * @param c      triangle vertex 3
-     * @param verts  list of vertices to test (excluding a,b,c)
-     * @param normal plane normal (used for projection)
-     * @return true if at least one vertex is inside the triangle
-     */
     private static boolean hasVerticesInside(ClipVertex a, ClipVertex b, ClipVertex c,
                                              List<ClipVertex> verts, Vec3d normal) {
-        // Project vertices onto the plane's 2D coordinate system.
         Vec3d u = normal.crossProduct(new Vec3d(1, 0, 0));
         if (u.length() < 1e-8) u = normal.crossProduct(new Vec3d(0, 1, 0));
         u = u.normalize();
@@ -520,19 +383,6 @@ public final class MeshPlaneCutter {
         return false;
     }
 
-    /**
-     * 2D point‑in‑triangle test using barycentric coordinates.
-     *
-     * @param px point x
-     * @param py point y
-     * @param ax triangle vertex 1 x
-     * @param ay triangle vertex 1 y
-     * @param bx triangle vertex 2 x
-     * @param by triangle vertex 2 y
-     * @param cx triangle vertex 3 x
-     * @param cy triangle vertex 3 y
-     * @return true if the point is inside the triangle (including edges)
-     */
     private static boolean pointInTriangle2D(double px, double py,
                                              double ax, double ay,
                                              double bx, double by,
@@ -545,17 +395,6 @@ public final class MeshPlaneCutter {
         return !(hasNeg && hasPos);
     }
 
-    /**
-     * Computes the signed area of the triangle (p1,p2,p3) in the XY plane.
-     *
-     * @param p1x point 1 x
-     * @param p1y point 1 y
-     * @param p2x point 2 x
-     * @param p2y point 2 y
-     * @param p3x point 3 x
-     * @param p3y point 3 y
-     * @return the signed area (2× area)
-     */
     private static double sign2D(double p1x, double p1y,
                                  double p2x, double p2y,
                                  double p3x, double p3y) {
